+++ srconvert.h
@@ -260,7 +260,7 @@ private:
 		f.numerator *= ci.overSamplingFactor;
 		f.denominator *= ci.overSamplingFactor;
 
-		FIRFilter<FloatType> firFilter(filterTaps.data(), filterTaps.size());
+		FIRFilter<FloatType> firFilter(filterTaps.data(), static_cast<int>(filterTaps.size()));
 		convertStages.emplace_back(f.numerator, f.denominator, firFilter, isBypassMode);
 		groupDelay = (ci.bMinPhase || !ci.bDelayTrim) ? 0 : (filterTaps.size() - 1) / 2 / f.denominator;
 		if (isBypassMode)
@@ -319,7 +319,7 @@ private:
 			std::vector<FloatType> filterTaps = makeFilterCoefficients<FloatType>(stageCi, fractions[i]);
 
 			// make the filter
-			FIRFilter<FloatType> firFilter(filterTaps.data(), filterTaps.size());
+			FIRFilter<FloatType> firFilter(filterTaps.data(), static_cast<int>(filterTaps.size()));
 
 			if (ci.bShowStages) { // dump stage parameters:
 				std::cout << "Stage: " << 1 + i << "\n";
